fix(chat): deleteRoom sockets params (#1446)
* fix(chat): deleteRoom sockets params

* fix(chat): deleteRoom sockets params pt2

* fix(chat): deleteRoom sockets params pt3 , null room early return

+++ modules/chat/src/Chat.ts
@@ -226,9 +226,22 @@ export default class Chat extends ManagedModule<Config> {
 
   async deleteRoom(call: GrpcRequest<DeleteRoomRequest>, callback: GrpcCallback<Room>) {
     const { _id } = call.request;
-
     let errorMessage: string | null = null;
-    const room: models.ChatRoom = await models.ChatRoom.getInstance()
+
+    const room = await models.ChatRoom.getInstance()
+      .findOne({ _id })
+      .catch((e: Error) => {
+        errorMessage = e.message;
+        return null;
+      });
+    if (!isNil(errorMessage)) {
+      return callback({ code: status.INTERNAL, message: errorMessage });
+    }
+    if (isNil(room)) {
+      return callback({ code: status.NOT_FOUND, message: 'Room not found' });
+    }
+
+    await models.ChatRoom.getInstance()
       .deleteOne({ _id })
       .catch((e: Error) => {
         errorMessage = e.message;
